requirements.txt
torch
torchvision
torchaudio
streamlit
Pillow
matplotlib
